On freedom of expression as a soon to be underground idea
I have a friend who I dearly respect and he believes that by mentioning
the name of a law, any law, we will be framed and hung for attempting to
resist. He is like many people in history: he is working to create
change and to build a better world that is far beyond anything I will
ever create in my lifetime. He believes that by removing some letters,
we'll be free to frame the debate, that we'll ensure that we still have
a voice if and when we find ourselves granted a moment to use it.

He believes that we still have a voice. My friend is an optimist.

This reminds me a great deal of another friend under very serious legal
charge. My friend and I have had lots of discussions about what it means
to be free. I mean of course the freedom that starts in your mind and of
course from that all other freedom follows. When one believes themselves
to be free, they act differently. When one believes themselves to be on
the edge, they make compromises. When one believes themselves to be
enslaved, they play the part. Not everyone fits that mold. My friend who
feels unfree expressed to me that they felt stress about not being able
to make one choice or another, that later after a choice has been made,
that they regretted the choices that they felt forced to take; I tried
my best to assure them that the worst is yet to come - this kind of
thinking leads to an absence of regret. This extremely oppressive
thought process forces one to suppress the choices they could have
freely made and after the fact, they didn't even realize it. Not even
long after the fact and certianly not as part of a process of regret.

This is a commit that depresses me but I will make it anyway. At least
I still have the freedom to feel that this is a choice, even if it is
one born of coercion.